+++ apps/backoffice-v2/src/lib/blocks/variants/DefaultBlocks/hooks/useDefaultBlocksLogic/useDefaultBlocksLogic.tsx
@@ -25,7 +25,6 @@ import { useKybRegistryInfoBlock } from '@/lib/blocks/hooks/useKybRegistryInfoBl
 import { useMainContactBlock } from '@/lib/blocks/hooks/useMainContactBlock/useMainContactBlock';
 import { useMainRepresentativeBlock } from '@/lib/blocks/hooks/useMainRepresentativeBlock/useMainRepresentativeBlock';
 import { useMapBlock } from '@/lib/blocks/hooks/useMapBlock/useMapBlock';
-import { useProcessTrackerBlock } from '@/lib/blocks/hooks/useProcessTrackerBlock/useProcessTrackerBlock';
 import { useProcessingDetailsBlock } from '@/lib/blocks/hooks/useProcessingDetailsBlock/useProcessingDetailsBlock';
 import { useRegistryInfoBlock } from '@/lib/blocks/hooks/useRegistryInfoBlock/useRegistryInfoBlock';
 import { useStoreInfoBlock } from '@/lib/blocks/hooks/useStoreInfoBlock/useStoreInfoBlock';
@@ -36,29 +35,25 @@ import { useWebsiteMonitoringBlock } from '@/lib/blocks/hooks/useWebsiteMonitori
 import { useCaseBlocks } from '@/lib/blocks/variants/DefaultBlocks/hooks/useCaseBlocksLogic/useCaseBlocks';
 import { useCaseDecision } from '@/pages/Entity/components/Case/hooks/useCaseDecision/useCaseDecision';
 import { useCaseState } from '@/pages/Entity/components/Case/hooks/useCaseState/useCaseState';
-import { omitPropsFromObject } from '@/pages/Entity/hooks/useEntityLogic/utils';
 import { selectDirectorsDocuments } from '@/pages/Entity/selectors/selectDirectorsDocuments';
 import { Send } from 'lucide-react';
-import { useCallback, useMemo } from 'react';
+import React, { useCallback, useMemo } from 'react';
 import { toast } from 'sonner';
 import { useCurrentCaseQuery } from '@/pages/Entity/hooks/useCurrentCaseQuery/useCurrentCaseQuery';
-import { useCasePlugins } from '@/pages/Entity/hooks/useCasePlugins/useCasePlugins';
-import { DEFAULT_PROCESS_TRACKER_PROCESSES } from '@/common/components/molecules/ProcessTracker/constants';
 import { useWebsiteMonitoringReportBlock } from '@/lib/blocks/variants/WebsiteMonitoringBlocks/hooks/useWebsiteMonitoringReportBlock/useWebsiteMonitoringReportBlock';
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 import { useAddressBlock } from '@/lib/blocks/hooks/useAddressBlock/useAddressBlock';
 import { getAddressDeep } from '@/pages/Entity/hooks/useEntityLogic/utils/get-address-deep/get-address-deep';
+import { useCaseOverviewBlock } from '@/lib/blocks/hooks/useCaseOverviewBlock/useCaseOverviewBlock';
+import { useSearchParamsByEntity } from '@/common/hooks/useSearchParamsByEntity/useSearchParamsByEntity';
+import { useLocation } from 'react-router-dom';
+import { omitPropsFromObjectWhitelist } from '@/common/utils/omit-props-from-object-whitelist/omit-props-from-object-whitelist';
 
-const pluginsOutputBlacklist = [
-  'companySanctions',
-  'directors',
-  'ubo',
-  'businessInformation',
-  'merchantMonitoring',
-  'website_monitoring',
-] as const;
+const registryInfoWhitelist = ['open_corporates'] as const;
 
 export const useDefaultBlocksLogic = () => {
+  const [{ activeTab }] = useSearchParamsByEntity();
+  const { search } = useLocation();
   const { data: workflow, isLoading } = useCurrentCaseQuery();
   const { data: session } = useAuthenticatedUserQuery();
   const caseState = useCaseState(session?.user, workflow);
@@ -136,12 +131,15 @@ export const useDefaultBlocksLogic = () => {
     childWorkflow => childWorkflow?.context?.entity?.type === 'business',
   );
 
-  const filteredPluginsOutput = useMemo(
-    () => omitPropsFromObject(workflow?.context?.pluginsOutput, ...pluginsOutputBlacklist),
-    [pluginsOutputBlacklist, workflow?.context?.pluginsOutput],
+  const registryInfo = useMemo(
+    () =>
+      omitPropsFromObjectWhitelist({
+        object: workflow?.context?.pluginsOutput,
+        whitelist: registryInfoWhitelist,
+      }),
+    [workflow?.context?.pluginsOutput],
   );
 
-  const pluginsOutputKeys = Object.keys(filteredPluginsOutput ?? {});
   const directorsDocuments = useMemo(() => selectDirectorsDocuments(workflow), [workflow]);
   const directorDocumentPages = useMemo(
     () =>
@@ -188,9 +186,9 @@ export const useDefaultBlocksLogic = () => {
   );
 
   const registryInfoBlock = useRegistryInfoBlock({
-    pluginsOutputKeys,
-    filteredPluginsOutput,
-    workflow,
+    registryInfo,
+    workflowId: workflow?.id,
+    documents: workflow?.context?.documents,
   });
 
   const kybRegistryInfoBlock = useKybRegistryInfoBlock({
@@ -242,7 +240,7 @@ export const useDefaultBlocksLogic = () => {
   });
 
   const mapBlock = useMapBlock({
-    address: getAddressDeep(filteredPluginsOutput, {
+    address: getAddressDeep(registryInfo, {
       propertyName: 'registeredAddressInFull',
     }),
     entityType: workflow?.context?.entity?.type,
@@ -337,13 +335,6 @@ export const useDefaultBlocksLogic = () => {
     [mutateEvent],
   );
 
-  const plugins = useCasePlugins({ workflow });
-  const processTrackerBlock = useProcessTrackerBlock({
-    workflow,
-    plugins,
-    processes: DEFAULT_PROCESS_TRACKER_PROCESSES,
-  });
-
   const associatedCompaniesBlock = useAssociatedCompaniesBlock({
     workflows: kybChildWorkflows,
     onClose,
@@ -398,6 +389,8 @@ export const useDefaultBlocksLogic = () => {
   const documentReviewBlocks = useDocumentReviewBlocks();
   const businessInformationBlocks = useKYCBusinessInformationBlock();
 
+  const caseOverviewBlock = useCaseOverviewBlock();
+
   const allBlocks = useMemo(() => {
     if (!workflow?.context?.entity) return [];
 
@@ -423,10 +416,10 @@ export const useDefaultBlocksLogic = () => {
       parentDocumentBlocks,
       associatedCompaniesBlock,
       associatedCompaniesInformationBlock,
-      processTrackerBlock,
       websiteMonitoringBlocks,
       documentReviewBlocks,
       businessInformationBlocks,
+      caseOverviewBlock,
     ];
   }, [
     associatedCompaniesBlock,
@@ -450,27 +443,32 @@ export const useDefaultBlocksLogic = () => {
     ubosRegistryProvidedBlock,
     websiteBasicRequirementBlock,
     websiteMonitoringBlock,
-    processTrackerBlock,
     websiteMonitoringBlocks,
     documentReviewBlocks,
     businessInformationBlocks,
+    caseOverviewBlock,
     workflow?.context?.entity,
   ]);
 
-  const {
-    activeTab,
-    blocks = [],
-    tabs,
-    setActiveTab,
-  } = useCaseBlocks({
+  const { blocks, tabs } = useCaseBlocks({
     workflow,
     config: workflow?.workflowDefinition?.config,
     blocks: allBlocks,
     onReuploadNeeded,
     isLoadingReuploadNeeded,
+    activeTab,
   });
-
   const availableTabs = useMemo(() => tabs.filter(tab => !tab.hidden), [tabs]);
+  const getUpdatedSearchParamsWithActiveTab = useCallback(
+    ({ tab }: { tab: string }) => {
+      const searchParams = new URLSearchParams(search);
+
+      searchParams.set('activeTab', tab);
+
+      return searchParams.toString();
+    },
+    [search],
+  );
 
   return {
     blocks,
@@ -478,7 +476,7 @@ export const useDefaultBlocksLogic = () => {
     isLoadingReuploadNeeded,
     isLoading,
     activeTab,
+    getUpdatedSearchParamsWithActiveTab,
     tabs: availableTabs,
-    setActiveTab,
   };
 };