+++ apps/backoffice-v2/src/pages/NotFound/NotFound.tsx
@@ -1,8 +1,10 @@
 import React, { FunctionComponent } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
+import { useRedirectToRootUrl } from '@/common/hooks/useRedirectToRootUrl/useRedirectToRootUrl';
 
 export const NotFoundRedirect: FunctionComponent = () => {
   const { state } = useLocation();
+  const urlToRoot = useRedirectToRootUrl();
 
-  return <Navigate to={'/en'} replace state={state} />;
+  return <Navigate to={urlToRoot} replace state={state} />;
 };