+++ services/workflows-service/src/workflow/hook-callback-handler.service.ts
@@ -33,29 +33,34 @@ export const ReportWithRiskScoreSchema = z
   })
   .passthrough();
 
-const setPluginStatusToSuccess = ({
+const removeLastKeyFromPath = (path: string) => {
+  return path?.split('.')?.slice(0, -1)?.join('.');
+};
+
+export const setPluginStatusToSuccess = ({
   resultDestinationPath,
   context,
   data,
+  ignoreLastKey = true,
 }: {
   resultDestinationPath: string;
   context: Record<string, unknown>;
   data: Record<string, unknown>;
+  ignoreLastKey?: boolean;
 }) => {
-  const removeLastKeyFromPath = (path: string) => {
-    return path?.split('.')?.slice(0, -1)?.join('.');
-  };
-
   const resultDestinationPathWithoutLastKey = removeLastKeyFromPath(resultDestinationPath);
-  const result = get(context, resultDestinationPathWithoutLastKey);
+  const result = get(
+    context,
+    ignoreLastKey ? resultDestinationPathWithoutLastKey : resultDestinationPath,
+  );
 
-  const resultWithData = set({}, resultDestinationPath, data);
+  const resultWithData = set({}, resultDestinationPath, ignoreLastKey ? data : { data });
 
   //@ts-ignore
   if (isObject(result) && result.status) {
     return set(
       resultWithData,
-      `${resultDestinationPathWithoutLastKey}.status`,
+      `${ignoreLastKey ? resultDestinationPathWithoutLastKey : resultDestinationPath}.status`,
       ProcessStatus.SUCCESS,
     );
   }
@@ -250,6 +255,7 @@ export class HookCallbackHandlerService {
       resultDestinationPath,
       context: workflowRuntime.context,
       data: reportData,
+      ignoreLastKey: false,
     });
   }