+++ apps/workflows-dashboard/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/workflows-dashboard
 
+## 0.2.5
+
+### Patch Changes
+
+- Bump
+
+## 0.2.4
+
+### Patch Changes
+
+- Bump
+
+## 0.2.3
+
+### Patch Changes
+
+- document changes
+
 ## 0.2.2
 
 ### Patch Changes